Add Model Predictive Control code
* Set timestamp length to 10 and dt between each
timestamp to 0.1
* Set reference velocity to 40mph
* Define the cost base on the reference state, the
actuators (delta, a) and the actuators change between
t+1 and t.
* Calculate the model using the given formulas.
* Fill MPC::Solve() which calculates the next
actuators values for delta and a. These values are
applied while the car is driving, then the state is
updated, the cost is calculated again and the next
actuators are computed; the process is repeated each
dt seconds.